requirements.txt
openai
openai-agents
arxiv
python-dotenv
pydantic
sib-api-v3-sdk
